requirements.txt

Flask==0.10.1
flask-cors==2.1.2
google-api-python-client==1.5.0
